Make the self-tests aware of our test harness
These changes make the self tests aware of our test harness, by scheduling a
recipe set for each harness.

Note, the function which determines if it's compatible with the distro is
currently just stringly typed. That's not a great comparison, but for the
foreseeable future we'll won't change the distro matrix heavily. So that shall
do for now.

https://projects.engineering.redhat.com/browse/BKR-3941

Change-Id: I3c309b72324223705f7f393401182aa4721e978e