remove all ia64 from the matrix
It is long-dead now and Restraint no longer works properly on it,
because of yum multilib versions conflicts.

Specifically, yum gets confused because the latest version of restraint
is not built for ia64 (ia64 has been removed entirely from Brew) but
older versions are, and thus it tries to install an older version of
restraint-rhts.ia64 with a newer version of restraint.i386 and then just
gives up.

Change-Id: Ib34434cf6fdfde831cae2fcb0c27563de496c370