let Beaker pick the default harness, even on RHEL8
With bug 1589610 implemented, Beaker should pick Restraint as the
default harness on RHEL8 and Fedora 29 without the user explicitly
requesting it, and we want to actually test that behaviour here too.

Refactored to remove the Harness type and values, since the script in
general doesn't need to actually care about the harness except for those
older distros where we support opting in to Restraint.